Started to consolidate with Marian's version by having Backend inherit from a Program instance (either NNLOJET or HEJ at the moment). A little hacky I think, but it could work so we're all on the same branch
