+++ src/runmodes.py
@@ -1,0 +1,5 @@
+import src.programs
+
+mode_selector = { "NNLOJET": src.programs.NNLOJET,
+                  "HEJ": src.programs.HEJ,
+}
